Move Person Computer conversation animation into a separate file
